requirements.txt
langchain==0.3.7
azure-search-documents==11.5.2
azure-identity==1.19.0
python-dotenv==1.0.1
streamlit==1.40.1
langchain-community==0.3.7
unstructured==0.16.5
langchain-openai==0.2.9